requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
torch==2.2.1
numpy==1.26.4
librosa==0.10.1
python-dotenv==1.0.1 